+++ src/DotPulsar/Abstractions/IValidateRemoteCertificate.cs
@@ -1,0 +1,29 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace DotPulsar.Abstractions;
+
+using System.Net.Security;
+using System.Security.Cryptography.X509Certificates;
+
+/// <summary>
+/// Abstraction for implementing remote certificate validation
+/// </summary>
+public interface IValidateRemoteCertificate
+{
+    /// <summary>
+    /// Callback triggered when creating an SslStream
+    /// </summary>
+    bool Validate(object sender, X509Certificate? certificate, X509Chain? chain, SslPolicyErrors sslPolicyErrors);
+}